[Oztechan/CCC#1929] Create :backend:app module
