+++ backend/app/src/main/kotlin/com/oztechan/ccc/backend/app/routes/RootRoute.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Mustafa Ozhan. All rights reserved.
  */
 
-package com.oztechan.ccc.backend.routes
+package com.oztechan.ccc.backend.app.routes
 
 import co.touchlab.kermit.Logger
 import io.ktor.http.ContentType
@@ -13,13 +13,13 @@ import io.ktor.server.response.respondText
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
 
-private const val PATH_ERROR = "/error"
-private const val ERROR_HTML = "error.html"
+private const val PATH_ROOT = "/"
+private const val INDEX_HTML = "index.html"
 
-internal suspend fun Route.getError() = get(PATH_ERROR) {
-    Logger.i { "GET Request $PATH_ERROR" }
+internal suspend fun Route.getRoot() = get(PATH_ROOT) {
+    Logger.i { "GET Request $PATH_ROOT" }
 
-    javaClass.classLoader?.getResource(ERROR_HTML)?.readText()?.let { resource ->
+    javaClass.classLoader?.getResource(INDEX_HTML)?.readText()?.let { resource ->
         call.respondText(
             text = resource,
             contentType = ContentType.Text.Html,
