+++ backend/app/src/main/resources/application.conf
@@ -1,0 +1,14 @@
+ktor {
+    deployment {
+        port = 8080
+        port = ${?PORT}
+    }
+    application {
+        modules = [
+            LoggerModuleKt.loggerModule,
+            KoinModuleKt.koinModule,
+            SyncModuleKt.syncModule,
+            APIModuleKt.apiModule
+        ]
+    }
+}
