--- backend/src/main/resources/application.conf
@@ -1,14 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-        port = ${?PORT}
-    }
-    application {
-        modules = [
-            com.oztechan.ccc.backend.module.LoggerModuleKt.loggerModule,
-            com.oztechan.ccc.backend.module.KoinModuleKt.koinModule,
-            com.oztechan.ccc.backend.module.SyncModuleKt.syncModule,
-            com.oztechan.ccc.backend.module.APIModuleKt.apiModule
-        ]
-    }
-}
